requirements.txt
asgiref==3.9.2
Django==5.2.7
django-tenants==3.9.0
pillow==11.3.0
psycopg2-binary==2.9.10
setuptools==80.9.0
sqlparse==0.5.3
wheel==0.45.1
django-environ==0.11.2 # Adicionado para gerenciamento de settings
whitenoise==6.6.0 # Adicionado para servir arquivos estáticos